+++ src/main/angular/src/app/app.module.ts
@@ -77,8 +77,8 @@ import {OrderComponent} from './order/order.component';
 import {PreSummaryComponent} from './admin/photo-report-editor/pre-summary/pre-summary.component';
 import {PreReportsGalleryComponent} from './admin/photo-report-editor/pre-reports-gallery/pre-reports-gallery.component';
 import {PreReportEditorComponent} from './admin/photo-report-editor/pre-report-editor/pre-report-editor.component';
-import {PageNotFoundComponent} from './page-not-found/page-not-found.component';
-import {TopBannerComponent} from './top-banner/top-banner.component';
+import {PageNotFoundComponent} from './pages/page-not-found/page-not-found.component';
+import {TopBannerComponent} from './components/top-banner/top-banner.component';
 import {StatisticService} from "./services/statistic/statistic.service";
 import {NgxUploaderModule} from "ngx-uploader";
 import {RulesComponent} from './pages/rules/rules.component';
