+++ src/main/angular/src/app/components/top-banner/top-banner.component.html
